- Fixed various bugs [ 1513150 ], [ 1513147 ], [ 1509571 ] (yalnifj)  - Fixed [ 1508850 ] User Admin Page Changes GEDCOMs (yalnifj)  - Fixed [ 1510437 ] slklist.php - EXCEL (SLK) - Errors, removed the file (yalnifj)  - Fixed [ 1509731 ] Delete Gedcom (yalnifj)
git-svn-id: svn://svn.webtrees.net/trunk@183 4201c862-f99b-1943-94d0-25fd05010cbc

+++ phpGedView/includes/functions_db.php
@@ -2398,9 +2398,15 @@ function find_rin_id($rin) {
 	return $rin;
 }
 
+/**
+ * Delete a gedcom from the database and the system
+ * Does not delete the file from the file system
+ * @param string $ged 	the filename of the gedcom to delete
+ */
 function delete_gedcom($ged) {
 	global $INDEX_DIRECTORY, $TBLPREFIX, $pgv_changes, $DBCONN, $GEDCOMS;
 
+	if (!isset($GEDCOMS[$ged])) return;
 	$dbged = $GEDCOMS[$ged]["id"];
 
 	$sql = "DELETE FROM ".$TBLPREFIX."blocks WHERE b_username='".$DBCONN->escapeSimple($ged)."'";
